Update the Grafana to support maxDataPoints and and intervalMs
Grafana queries include two hints: maxDataPoints, and intervalMs.
maxDataPoints indicates how many pixesl are actually in the window
of the display pane. intervalMs is (roughly) the time-period in
milliseconds for each of those samples. These hints are now passed
along to the Sos as_ndarray() function to interpolate/extrapolate
as necessary for better presentation.

This change requires the latest SOS that includes support for
interpolation in the Filter.as_ndarray() method.

The implementation of the back-end logic in models_sos and views
were not using the Numpy array views support that allows you to
change the 'look' of an array actually copying the data. Also, the
query code was replicated across the different models. The logic was
refactored to collapse the query code into a single parent-class
function and the array manipulation was modified to use the numpy
ndarray views support.

The test analysis for Log and Linear were removed as these will
not ever be used.